WoT restructure A6: stickyScrollStrategy.ts -> scroll/ (git-mv + imports)
