+++ handsontable/src/3rdparty/walkontable/src/scroll/stickyScrollStrategy.ts
@@ -1,7 +1,7 @@
-import { isSafari } from '../../../helpers/browser';
-import type { EngineContext } from './wire';
-import type { default as Overlays } from './overlays';
-import type { Overlay } from './overlay/_base';
+import { isSafari } from '../../../../helpers/browser';
+import type { EngineContext } from '../wire';
+import type { default as Overlays } from '../overlays';
+import type { Overlay } from '../overlay/_base';
 
 /**
  * Assembles the StickyScrollStrategy's dependencies. Most come from the engine composition context;
